Drillo/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
asteval==0.9.31
astunparse==1.6.3
attrs==23.1.0
autopep8==2.0.4
boto3==1.28.27
botocore==1.31.27
bw-migrations==0.2
bw2calc==1.8.2
bw2data==3.6.6
bw2io==0.8.7
bw2parameters==1.1.0
cached-property==1.5.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
constructive_geometries==0.8.2
country-converter==1.0.0
cryptography==41.0.4
datapackage==1.15.2
diff-match-patch==20230430
Django==4.2.4
django-cors-headers==4.2.0
django-dynamic-model==0.3.0
django-dynamic-models==0.0.5
django-import-export==4.1.1
django-storages==1.14.1
djangorestframework==3.14.0
docopt==0.6.2
eight==1.0.1
et-xmlfile==1.1.0
fasteners==0.18
future==0.18.3
greenlet==2.0.2
idna==3.4
ijson==3.2.3
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
jsonlines==3.1.0
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
linear-tsv==1.1.0
llvmlite==0.40.1
lxml==4.9.3
MarkupSafe==2.1.3
mrio-common-metadata==0.2.1
numba==0.57.1
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
peewee==3.16.3
Pint==0.22
prettytable==3.8.0
psutil==5.9.5
pycodestyle==2.11.0
pycparser==2.21
PyMySQL==1.1.0
PyPrind==2.11.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pyxlsb==1.0.10
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rfc3986==2.0.0
rpds-py==0.9.2
s3transfer==0.6.2
scipy==1.11.1
six==1.16.0
sparse==0.14.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
stats-arrays==0.6.5
tableschema==1.20.2
tablib==3.5.0
tabulator==1.53.5
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
unfold==1.0.8
unicodecsv==0.14.1
Unidecode==1.3.6
urllib3==1.26.16
voluptuous==0.13.1
wcwidth==0.2.6
Whoosh==2.7.4
wrapt==1.15.0
wurst==0.3.4
xlrd==2.0.1
XlsxWriter==3.1.2
